for #660: split asserts.token
